+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 
 ### Bugfix
 * fix loading of configuration inside the `AutoRuleCorpusTester` for `logprep test integration`
+* fix auto rule tester (`test unit`), which was broken after adding support for multiple configuration files and resolving paths in configuration files 
 
 ## v10.0.2
 ### Bugfix