+++ CHANGELOG.md
@@ -8,6 +8,9 @@
 ### Improvements
 
 ### Bugfix
+
+## v10.0.3
+### Bugfix
 * fix loading of configuration inside the `AutoRuleCorpusTester` for `logprep test integration`
 * fix auto rule tester (`test unit`), which was broken after adding support for multiple configuration files and resolving paths in configuration files